feat: extension of backend interface to display real peaks in frontend

+++ cmd/mb3server/src/model_mb_record_peak_peak_values_inner.go
@@ -13,13 +13,13 @@ package mb3server
 type MbRecordPeakPeakValuesInner struct {
 
 	// m/z of the peak
-	Mz float64 `json:"mz,omitempty"`
+	Mz float32 `json:"mz,omitempty"`
 
 	// intensity of the peak
 	Intensity float64 `json:"intensity,omitempty"`
 
 	// Relative intensity of the peak
-	Rel int32 `json:"rel,omitempty"`
+	Rel uint16 `json:"rel,omitempty"`
 }
 
 // AssertMbRecordPeakPeakValuesInnerRequired checks if the required fields are not zero-ed
